Chapter 10 -- Functions: Closures
So it turns out some objects don't like to derive PartialEq. Fair
enough.

One of those types is Environment. Also I had to decide what it means
for two functions to be equal. I think the most obvious thing to do is
wrap functions in `Rc`s (they're immutable anyways) and compare equal if
the Rcs are equal.

To avoid manually impling everything for Object, for the sake of just
two variants, I made a newtype representing Rc<function data> and that
seems to be decently ergonomic.