Chapter 17 -- Compiling Expressions
Since I'm staying in the same language, I don't have to re-implement the
parser. I'm just going to take the AST as input to the compiler.

On one hand, it'd be more efficient to not build the whole AST. On the
other, I would rather quit than re-implement the whole-ass parser.