Chapter 19 -- Strings
This needs to be touched up a bit. In particular, I need to add a list
of objects to the VM. The tricky part is I'm not sure how to track
ownership there.

I think the VM should own the objects, but objects need to reference
each other, potentially mutably. I can't use "index in the list", as
that will shift with garbage collection. Maybe make the object pool a
map from IDs to objects, and objects only store object IDs?