[WIP] Garbage Collection
Implemented a mark-and-sweep garbage collector. (TODO add a test that
rooted handles work as expected. also, improve the traceable trait, and
make downgrades more ergonomic pls).

In addition, I had to tweak a few things.

In order to get constants with heap-allocated memory (e.g. constant
strings) to stick around, the compiler needs access to the VM heap, so
it can put the objects in there, and the chunks need some strong GC
pointers, so that the objects don't get garbage collected.

Also, the format of Value changed a bit (still not totally settled on
implementation), and I moved `try_add` to the VM file, since it might
need to make a heap allocation.