Split value types
The kinds of values that can be stored in a chunk are subtly different
from those that are allowed, live, at runtime. For example, there's no
possibility of reference cycles, so there's no GcPtrs anywhere.

But more importantly, this sets us up cleanly for closures -- function
templates are chunk-only values, and when loaded, turn into real live
closures, with upvalues and all.