Chapter 25 -- Closures (2)
Okay, we have upvalues now. Getting the type right was tricky, but I
think this'll do it. We want to be able to share upvalue lists cheaply
(consider the case of a recursive function), and also to be able to
mutate an upvalue when it comes time to close it.

Still can't close them yet though.