Split heap into multiple heaps
It's kind of frustrating to store LoxClosures behind a Gc<HeapObject>,
and then have to do a whole song and dance to retrieve it. So instead of
having a heap of HeapObjects, we eliminate the concept of HeapObjects
entirely, and just have GcPtrs of LoxClosures, LoxInstances, etc.

The upside is that Value becomes a flat enum; no more nesting. The
downside is now we have many more heaps to manage, and I think this kills
any dream I had about having an incremental garbage collection.

But I think this is nicer overall. Some macro nonsense may be needed to
make things even better.