Change lexer-parser relationship
I wanted to combine the lexing and parsing phase into one thing, so that
a lexer error doesn't inhibit additional parser errors from being
created.

To do this, the lexer now returns guaranteed tokens when asked: no more
Nones and no more Errs. When the lexer cannot understand something, it
emits a Token::Error, and when the end of the stream is reached, it
returns Token::EndOfFile forever.

The parser now contains a lexer, not an iterator of tokens. I don't
think anyone is going to have a token stream lying around to hand to the
parser, so the parser should take just a source string as input.

At some point in the future, I'll do the same thing with the
interpreter and compiler -- they should take source strings, not ASTs.