Change how we do synchronization
Instead of re-syncing in parse_all(), we instead do it in
parse_spanned_declaration. This lets use recover from errors that are
not at the top level (i.e., within a function).

This means we have to store errors within parser state itself (instead
of a local Vec).

Lastly, this means parse_braced_statement_list changes a bit, so that we
can handle missing statements.