Fix remaining two tests
These both come from lexing errors, and so the only difficult bit was
"what if the first token is a bad one?". Since we dropped the result of
self.bump(), we lost the error message from the lexer.

To fix that, I moved the first bump from the constructor to `parse_all`.
This required changing the simple unit tests I had, but that's not too
bad (and also those tests are kind of obsolete now).

I also changed the error message for unexpected character to, you know,
include the character.