Add For variant to AST
It seems wrong to do this weird rebuilding in the parser. Maybe as a
lowering stage it makes sense (e.g. what I did in the resolver?). But
the parse tree should be mostly faithful.

Amazingly, this worked on the second try. Misplaced that last end_scope
at first, but it worked great after that.